% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_by_weekhour}
\alias{aggregate_by_weekhour}
\title{Aggregate by hour of the week}
\usage{
aggregate_by_weekhour(data, week_start = 1)
}
\arguments{
\item{data}{object of class \code{dockless_dfc}.}

\item{week_start}{day on which week starts following ISO conventions -
1 means Monday, 7 means Sunday.}
}
\value{
Returns a numeric value between 0 and 167.
}
\description{
Aggregates the distance column in each \code{dockless_df} data frame of a
\code{dockless_dfc} by the hour of the week.
}
