% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_loop.R
\name{build_models}
\alias{build_models}
\title{Forecasting models for \code{dockless_dfc} object}
\usage{
build_models(data, auto_seasonality = TRUE, seasons = list(NULL, 96,
  672, c(96, 672)))
}
\arguments{
\item{data}{object of class \code{dockless_dfc}.}

\item{auto_seasonality}{logical. If \code{TRUE}, the seasonal period of each
time series will be automatically determined with time series cross-validation.}

\item{seasons}{if \code{auto_seasonality} is set to \code{TRUE}: a list of
different seasonal period lengths - or, in the case of multiple seasonality,
combinations of seasonal period lenghts - in time lags to be tested for in
the time series cross-validation. No seasonality should be represented by
\code{NULL}. If \code{auto_seasonality} is set to \code{FALSE}: a list of
seasonal period lengths - or, in the case of multiple seasonality,
combinations of seasonal period lenghts - in time lags to be used for each
of the time series. No seasonality should be represented by \code{NULL}. The
list should be of the same length as the number of time series in the
\code{dockless_dfc} object.}
}
\value{
Returns a list of models, one for each \code{dockless_df} time series.
Each model is an object of class \code{ARIMA} for non-seasonal data and an
object of class \code{stlm} for seasonal data. Both classes come from the
\code{forecast} package.
}
\description{
Fits a forecasting model on each \code{dockless_df} time series in a
\code{dockless_dfc}. For non-seasonal time series, an ARIMA model will
be fitted on the log transformed data. For seasonal time series, the log
transformed data will be decomposed with STL decomposition, and an ARIMA
model will be fitted on the seasonally adjusted data.
}
