% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_loop.R
\name{build_single_model}
\alias{build_single_model}
\title{Forecasting model for \code{dockless_df} object}
\usage{
build_single_model(data, auto_seasonality = TRUE, seasons = list(NULL,
  96, 672, c(96, 672)))
}
\arguments{
\item{data}{object of class \code{dockless_df}.}

\item{auto_seasonality}{logical. If \code{TRUE}, the seasonal periods of the
time series will be automatically determined with time series cross-validation.}

\item{seasons}{if \code{auto_seasonality} is set to \code{TRUE}: a list of
different seasonal period lengths - or, in the case of multiple seasonality,
combinations of seasonal periods lenghts - in time lags to be tested for in
the time series cross-validation. No seasonality should be represented by
\code{NULL}. If \code{auto_seasonality} is set to \code{FALSE}: the seasonal
period length - or, in the case of multiple seasonality, combination of
seasonal period lenghts - in time lags. No seasonality should be represented
by \code{NULL}.}
}
\value{
Returns an object of class \code{ARIMA} for non-seasonal data and an
object of class \code{stlm} for seasonal data. Both classes come from the
\code{forecast} package.
}
\description{
Fits a forecasting model on a \code{dockless_df} time series. If the time
series is non-seasonal, an ARIMA model will be fitted on the log transformed
data. If the time series is seasonal, the log transformed data will be
decomposed with STL decomposition, and an ARIMA model will be fitted on the
seasonally adjusted data.
}
