% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fun.R
\name{choose_alpha_plot}
\alias{choose_alpha_plot}
\title{Homogeneity plot of spatially constrained hierarchical clustering of 
a \code{dockless_dfc} object}
\usage{
choose_alpha_plot(data, polygons, k, plot = TRUE, range_alpha = seq(0,
  1, 0.1))
}
\arguments{
\item{data}{object of class \code{dockless_dfc}.}

\item{polygons}{object of class \code{sf} with polygon geometry, in which each
feature refers to the spatial location of each \code{dockless_df} object in
the \code{dockless_dfc}.}

\item{k}{number of clusters.}

\item{plot}{logical; defining if the output should be a plot.}

\item{range_alpha}{the values of alpha to be tested for. Should be a vector
of real values between 0 and 1.}
}
\value{
If \code{plot} is \code{TRUE}, it returns a plot. Otherwise, it
returns an object of class \code{choicealpha} from the package \code{ClustGeo}.
}
\description{
The homogeneity plot is meant to choose the optimal value of the mixing 
parameter alpha, which sets the strength of the spatial constraint in
a spatially-constrained hierarchical clustering.
}
