% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fun.R
\name{choose_k_plot}
\alias{choose_k_plot}
\title{Dendrogram of non-spatial hierarchical clustering of a \code{dockless_dfc} object}
\usage{
choose_k_plot(data)
}
\arguments{
\item{data}{object of class \code{dockless_dfc}.}
}
\value{
Returns a plot.
}
\description{
The non-spatial dendrogram is meant to choose the number of clusters k used in
a spatially-constrained hierarchical clustering.
}
