% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_loop.R
\name{create_modelpoints}
\alias{create_modelpoints}
\title{Create model points}
\usage{
create_modelpoints(centroids)
}
\arguments{
\item{centroids}{all grid cell centroids as an \code{sf} object with point geometry,
containing at least the attributes \code{cluster}, specifying to which cluster each
grid cell centroid belongs, and \code{intensity}, specifying the number of pick-ups
in the grid cell that corresponds to the grid cell centroid.}
}
\value{
Returns an object of class \code{sf} with point geometry.
}
\description{
Creates an object of class \code{sf} containing the geographical locations
for which the forecasting models will be build. The locations are calculated
by taking, per cluster, the centroid of the grid cell centers , weighted by
the usage intensity of the grid cell polygons.
}
