% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_loop.R
\name{dissimilarity_data}
\alias{dissimilarity_data}
\title{Dissimilarity matrix of a \code{dockless_dfc} object}
\usage{
dissimilarity_data(data)
}
\arguments{
\item{data}{object of class \code{dockless_dfc}.}
}
\value{
Returns an object of class \code{dist}.
}
\description{
Creates a dissimilarity matrix based on the euclidean distances between all
\code{dockless_df} objects in a \code{dockless_dfc}.
}
