% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dissimilarity_spatial}
\alias{dissimilarity_spatial}
\title{Dissimilarity matrix of a \code{sf} object}
\usage{
dissimilarity_spatial(polygons)
}
\arguments{
\item{polygons}{object of class \code{sf} with polygon geometry.}
}
\value{
Returns an object of class \code{dist}.
}
\description{
Creates a dissimilarity matrix based on the adjacency of all polygons in a
\code{sf} object with polygon geometry.
}
