% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancedata_centroids.R
\docType{data}
\name{distancedata_centroids}
\alias{distancedata_centroids}
\title{Distance data for the grid centroids}
\format{An object of class dockless_dfc, containing 249 data frames of
subclass dockless_df, with 2688 rows.
\describe{
  \item{bike_id}{unique identification number of the bike}
  \item{time}{timestamp of the observation}
  \item{distance}{distance to the neareast available bike, in meters}
}}
\usage{
distancedata_centroids
}
\description{
The distance data for the centroids of the grid cells, during the
training period. Used for clustering.
}
\keyword{datasets}
