% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancedata_modelpoints.R
\docType{data}
\name{distancedata_modelpoints}
\alias{distancedata_modelpoints}
\title{Distance data for the model points}
\format{An object of class dockless_dfc, containing 4 data frames of
subclass dockless_df, with 2689 rows.
\describe{
  \item{bike_id}{unique identification number of the bike}
  \item{time}{timestamp of the observation}
  \item{distance}{distance to the neareast available bike, in meters}
}}
\usage{
distancedata_modelpoints
}
\description{
The distance data for the model points, during the
training period. Used for model building.
}
\keyword{datasets}
