% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancedata_modelpoints_test.R
\docType{data}
\name{distancedata_modelpoints_test}
\alias{distancedata_modelpoints_test}
\title{Distance data for the model points during the test period}
\format{An object of class dockless_dfc, containing 4 data frames of
subclass dockless_df, with 2117 rows.
\describe{
  \item{bike_id}{unique identification number of the bike}
  \item{time}{timestamp of the observation}
  \item{distance}{distance to the neareast available bike, in meters}
}}
\usage{
distancedata_modelpoints_test
}
\description{
The distance data for the model points, during the
test period. Used to create the plot in section 5.4.1.
}
\keyword{datasets}
