% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancedata_testpoints.R
\docType{data}
\name{distancedata_testpoints}
\alias{distancedata_testpoints}
\title{Distance data for the test points}
\format{An object of class dockless_dfc, containing 500 data frames of
subclass dockless_df, with 2117 rows.
\describe{
  \item{bike_id}{unique identification number of the bike}
  \item{time}{timestamp of the observation}
  \item{distance}{distance to the neareast available bike, in meters}
}}
\usage{
distancedata_testpoints
}
\description{
The distance data for the test points, both for decomposition, as for
the validation of the forecast results.
}
\keyword{datasets}
