% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_fun.R
\name{dockless_colors}
\alias{dockless_colors}
\title{Color schemes for the \code{dockless} package}
\usage{
dockless_colors(n = 10, categorical = FALSE)
}
\arguments{
\item{n}{the number of colors to be in the palette. Ignored if 
\code{categorical} is set to \code{TRUE}.}

\item{categorical}{logical. If \code{TRUE}, a categorical color scheme will be
produced. If \code{FALSE}, a sequential color scheme will be produced.}
}
\value{
Returns an vector of colors.
}
\description{
Creates either sequential or categorical color schemes in the dockless style.
}
