% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{error}
\alias{error}
\alias{error.dockless_fc}
\alias{error.dockless_fcc}
\title{Forecast error}
\usage{
error(x, type, average = TRUE)

\method{error}{dockless_fc}(x, type, average = TRUE)

\method{error}{dockless_fcc}(x, type, average = TRUE)
}
\arguments{
\item{x}{object either of class \code{dockless_fc} or \code{dockless_fcc}.}

\item{type}{one of 'RMSE' or 'MAE'.}

\item{average}{logical; if \code{TRUE}, errors for all \code{dockless_fc}
objects in a \code{dockless_fcc} are averaged. If \code{FALSE}, errors
for all \code{dockless_fc} objects in a \code{dockless_fcc} are given
seperately, in a numeric vector. Ignored if \code{x} is an object of class
\code{dockless_fc}.}
}
\value{
Returns either a numeric value or a vector of numeric values.
}
\description{
Calculates either the root mean squared error or the mean absolute
error of a set of forecasts.
}
