% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{error_hourofday}
\alias{error_hourofday}
\alias{error_hourofday.dockless_fc}
\alias{error_hourofday.dockless_fcc}
\title{Forecast error per hour of the day}
\usage{
error_hourofday(x, type)

\method{error_hourofday}{dockless_fc}(x, type)

\method{error_hourofday}{dockless_fcc}(x, type)
}
\arguments{
\item{x}{object either of class \code{dockless_fc} or \code{dockless_fcc}..}

\item{type}{one of 'RMSE' or 'MAE'.}
}
\value{
Returns a vector of numeric values.
}
\description{
Calculates eiter the root mean squared error or the mean absolute
error, per hour of the day.
}
