% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate a \code{dockless_fcc} object}
\usage{
evaluate(x, clusters, type)
}
\arguments{
\item{x}{object either of class \code{dockless_fcc}.}

\item{clusters}{vector specifying to which cluster each test point belongs.}

\item{type}{one of 'RMSE' or 'MAE'.}
}
\value{
Returns a matrix.
}
\description{
Calculates the average, maximum and minimum root mean squared error or
mean absolute of a \code{dockless_fcc} object, both for the total area
as per cluster.
}
