% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_fun.R
\name{forecast_day}
\alias{forecast_day}
\title{Forecast one day ahead}
\usage{
forecast_day(location, time, clusters = NULL, models = NULL,
  naive = FALSE, weeks_of_data = 8, database_user, database_password)
}
\arguments{
\item{location}{object of class \code{sf}, \code{sfc} or \code{sfg}, with
point geometry.}

\item{time}{timestamp of class \code{POSIXct}.}

\item{clusters}{list of objects of class \code{sf}, \code{sfc} or \code{sfg},
with polygon geometry, each representing the geographical outline of a cluster.}

\item{models}{list of objects of either class \code{ARIMA} or class \code{stlm},
each representing the forecasting model belonging to the corresponding cluster.}

\item{naive}{logical; if set to TRUE, only naive forecasts are made and the
parameters \code{clusters} and \code{models} are ignored.}

\item{weeks_of_data}{how many weeks of historical data to use for forecasting.}

\item{database_user}{character defining the user name to access the database.}

\item{database_password}{character defining the password to access the database.}
}
\value{
Returns an object of class \code{forecast}, coming from the
\code{forecast} package.
}
\description{
For a given location and a given time, forecast all the future timestamps
- i.e. every 15min - until one day ahead. Database credentials are needed
for this function.
}
