% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{forecast_evaluation}
\alias{forecast_evaluation}
\title{Evaluate a \code{dockless_fc_dfc} object}
\usage{
forecast_evaluation(x)
}
\arguments{
\item{x}{object of class \code{dockless_fc_dfc}.}
}
\value{
Returns a list.
}
\description{
Calculates the mean absolute error of a \code{dockless_fc_dfc} object, the
mean absolute error of each of the forecasted points in that object and the
mean absolute error of each day-of-the-week in that object.
}
