% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_fun.R
\name{forecast_lastday}
\alias{forecast_lastday}
\title{Forecast last day of a \code{dockless_df} object}
\usage{
forecast_lastday(data, model, weeks_of_data = NULL)
}
\arguments{
\item{data}{object of class \code{dockless_df}.}

\item{model}{forecasting model. Should be either of class \code{ARIMA}, for
non-seasonal data, or class \code{stlm}, for seasonal data.}

\item{weeks_of_data}{how many weeks of historical data to use for forecasting.
If set to \code{NULL}, all available data is used.}
}
\value{
Returns an object of class \code{dockless_fc_df}, which is a data
frame containing the observed values, forecasted values and corresponding
prediction intervals.
}
\description{
Forecasts the last day of a \code{dockless_df} time series, using a model
generated with either \link{build_single_model} or \link{build_models}.
}
