% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_loop.R
\name{forecast_lastweek}
\alias{forecast_lastweek}
\title{Forecast last week of a \code{dockless_df} object}
\usage{
forecast_lastweek(data, method, model = NULL)
}
\arguments{
\item{data}{object of class \code{dockless_df}.}

\item{method}{one of 'DBAFS', 'NFS' or 'EFS', specifying the forecasting method
to be used.}

\item{model}{forecasting model, either of class \code{ARIMA}, for
non-seasonal data, or class \code{stlm}, for seasonal data. Ignored if \code{method}
is set to either 'NFS' or 'EFS'.}
}
\value{
Returns an object of class \code{dockless_fc_df}, which is a data
frame containing the observed values, forecasted values and corresponding
prediction intervals.
}
\description{
Forecasts the last week of a \code{dockless_df} time series, using a model
generated with either \link{build_single_model} or \link{build_models}.
Each day will be forecasted seperately, with a period of data of given length.
The period of data to forecast with will always end at the start of the day
that is forecasted.
}
