% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_loop.R
\name{forecast_multiple}
\alias{forecast_multiple}
\title{Forecast a \code{dockless_dfc} object}
\usage{
forecast_multiple(data, method, perspective = "user", points,
  models = NULL)
}
\arguments{
\item{data}{object of class \code{dockless_dfc}.}

\item{method}{one of 'DBAFS', 'NFS' or 'EFS', specifying the forecasting method
to be used.}

\item{perspective}{one of 'user' or 'operator', specifying from which perspective
forecasts should be made.}

\item{points}{if \code{perspective} is set to 'user', the test points as object of
class \code{sf} with point geometry, obtained with the \code{create_testpoints}
function. If \code{perspective} is set to 'operator', the grid cell centroids as
object of class \code{sf} with point geometry, obtained with the \code{create_grid}
function.}

\item{models}{list of objects of either class \code{ARIMA} or class \code{stlm},
each representing the forecasting model belonging to a cluster. Must be obtained
with the \code{build_models} function. Ignored if \code{method} is set to either
'NFS' or 'EFS'.}
}
\value{
Returns an object of class \code{dockless_fcc}, which is a collection of
\code{dockless_fc} data frames.
}
\description{
Forecast a \code{dockless_dfc} object, using either DBAFS, NFS or EFS, from either
the user perspective or the operator perspective.
}
