% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_fun.R
\name{forecast_period}
\alias{forecast_period}
\title{Forecast a \code{dockless_dfc} object for a given period}
\usage{
forecast_period(data, clusters, models, length, last_day,
  weeks_of_data = 8)
}
\arguments{
\item{data}{object of class \code{dockless_dfc}.}

\item{clusters}{vector specifying to which cluster each \code{dockless_df}
time series belongs.}

\item{models}{list of forecasting models, with each model corresponding to one
of the clusters. Models Should be either of class \code{ARIMA}, for non-seasonal
data, or class \code{stlm}, for seasonal data.}

\item{length}{length in weeks of the forecasting period.}

\item{last_day}{character specifying the last day of the forecasting period.
Should be in the format 'YYYY-mm-dd'.}

\item{weeks_of_data}{how many weeks of historical data to use for forecasting.
If set to \code{NULL}, all available data is used.}
}
\value{
Returns an object of class \code{dockless_fc_dfc}, which is a list
containing several  - one for each week - matrices. Each matrix consists of
seven rows - one for each day - and a number of columns equal to the number
of forecasted \code{dockless_df} time series. Each element in such a matrix
is an object of class \code{dockless_fc_df}, which is a data frame containing
the observed values, forecasted values and corresponding prediction intervals
of a one-day forecast for one single time series.
}
\description{
Forecasts a given number of weeks for all \code{dockless_df} time series in
a \code{dockless_dfc}, using a set of models generated with \link{build_models}.
Each day will be forecasted seperately, with a period of data of given length.
The period of data to forecast with will always end at the start of the day
that is forecasted. Which of the models to use for each time series depends
on the cluster in which the corresponding forecast point is located.
}
