% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_loop.R
\name{forecast_single}
\alias{forecast_single}
\title{Forecast a \code{dockless_df} object}
\usage{
forecast_single(forecast_data, evaluate_data, method, model = NULL)
}
\arguments{
\item{forecast_data}{time series to be forecasted as object of
class \code{dockless_df}.}

\item{evaluate_data}{true observations of forecasted day, as
object of class \code{dockless_df}}

\item{method}{one of 'DBAFS', 'NFS' or 'EFS', specifying the forecasting method
to be used.}

\item{model}{forecasting model, either of class \code{ARIMA}, for
non-seasonal data, or class \code{stlm}, for seasonal data. Ignored if \code{method}
is set to either 'NFS' or 'EFS'.}
}
\value{
Returns an object of class \code{dockless_fc}, which is a data
frame containing the observed values, forecasted values and corresponding
prediction intervals.
}
\description{
Forecasts a \code{dockless_df} time series one day ahead.
}
