% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_loop.R
\name{forecast_testset}
\alias{forecast_testset}
\title{Forecast a test set}
\usage{
forecast_testset(intensity_data, first_day, n_weeks = 2,
  n_sample = 500, clusters = NULL, models = NULL, naive = FALSE,
  weeks_of_data = 8, database_user, database_password)
}
\arguments{
\item{intensity_data}{object of class \code{sf}, containing the locations and times
of all pick-ups in the bike sharing system.}

\item{first_day}{character specifying the first day of the test set period.
Should be in the format 'YYYY-mm-dd'.}

\item{n_weeks}{the length in weeks of the test set period.}

\item{n_sample}{the size of the sample for each week.}

\item{clusters}{list of objects of class \code{sf}, \code{sfc} or \code{sfg},
with polygon geometry, each representing the geographical outline of a cluster.}

\item{models}{list of objects of either class \code{ARIMA} or class \code{stlm},
each representing the forecasting model belonging to the corresponding cluster.}

\item{naive}{logical; if set to TRUE, only naive forecasts are made and the
parameters \code{clusters} and \code{models} are ignored.}

\item{weeks_of_data}{how many weeks of historical data to use for forecasting.}

\item{database_user}{character defining the user name to access the database.}

\item{database_password}{character defining the password to access the database.}
}
\value{
Returns an object of class \code{forecast}, coming from the
\code{forecast} package.
}
\description{
For all 'location, time'-combinations in a test set, forecast all the future
timestamps - i.e. every 15min - until one day ahead. The test set is created inside
the function by sampling 'location, time of the day, day of the week'-combinations
from a dataframe with all the pick ups in the historical data from the bike sharing
system. The days of the week from those combinations are translated into the same
days of the week in the test set period. Database credentials are needed for this
function.
}
