% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{mae}
\alias{mae}
\alias{mae.dockless_fc}
\alias{mae.dockless_fcc}
\title{Mean absolute error}
\usage{
mae(x)

\method{mae}{dockless_fc}(x)

\method{mae}{dockless_fcc}(x)
}
\arguments{
\item{x}{object either of class \code{dockless_fc} or \code{dockless_fcc}.}
}
\value{
Returns a numeric value.
}
\description{
Calculates the mean absolute error of one or more forecasts.
}
