% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_fun.R
\name{mae_days}
\alias{mae_days}
\title{Mean absolute error per day of the week}
\usage{
mae_days(x)
}
\arguments{
\item{x}{object of class \code{dockless_fc_dfc}.}
}
\value{
Returns a vector of numeric values.
}
\description{
Calculates the mean absolute error of each of day-of-the-weeks in a
\code{dockless_fc_dfc} object.
}
