% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{mae_lag}
\alias{mae_lag}
\alias{mae_lag.dockless_fc}
\alias{mae_lag.dockless_fcc}
\title{Mean absolute error per forecast lag}
\usage{
mae_lag(x)

\method{mae_lag}{dockless_fc}(x)

\method{mae_lag}{dockless_fcc}(x)
}
\arguments{
\item{x}{object either of class \code{dockless_fc} or \code{dockless_fcc}..}
}
\value{
Returns a vector of numeric values.
}
\description{
Calculates the mean absolute error per forecast lag.
}
