% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{mae_points}
\alias{mae_points}
\title{Mean absolute error per forecasting point}
\usage{
mae_points(x)
}
\arguments{
\item{x}{object of class \code{dockless_fc_dfc}.}
}
\value{
Returns a vector of numeric values.
}
\description{
Calculates the mean absolute error of each of the forecasted point in a
\code{dockless_fc_dfc} object.
}
