% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_fun.R
\name{mae_timeofday}
\alias{mae_timeofday}
\title{Mean absolute error per time of the day}
\usage{
mae_timeofday(x)
}
\arguments{
\item{x}{object of class \code{dockless_fc_dfc}.}
}
\value{
Returns a vector of numeric values.
}
\description{
Calculates the mean absolute error of each timestamp in a
\code{dockless_fc_dfc} object.
}
