% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_fun.R
\name{make_grid}
\alias{make_grid}
\title{Square grid over service area}
\usage{
make_grid(area, type, ...)
}
\arguments{
\item{area}{object of class \code{sf} representing the service area.}

\item{type}{either \code{polygons} or \code{centers}.}

\item{...}{further arguments passed to \code{st_make_grid}.}
}
\value{
If \code{type} is set to \code{polygons}, it returns an object 
of class \code{sf} containing the grid cells as square polygons.
If \code{type} is set to \code{centers} it returns an object of 
class \code{sf} containing the grid cell centroids as points.
}
\description{
Creates a square grid with given cell size over the service area 
of the dockless bike sharing system.
}
