% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{query_bikes}
\alias{query_bikes}
\title{Query unique bike id's}
\usage{
query_bikes(from = as.POSIXct("2018-09-10 16:30:00", format =
  "\%Y-\%m-\%d \%H:\%M:\%S", tz = "America/Los_Angeles"),
  to = Sys.time(), database_user, database_password)
}
\arguments{
\item{from}{timestamp of class \code{POSIXct} defining from which timestamp
on data should be queried. Default is set to the first timestamp
in the database (i.e. earlier timestamps than the default are not possible).}

\item{to}{timestamp of class \code{POSIXct} defining until which timestamp
data should be queried. Default is set to the system time.}

\item{database_user}{character defining the user name to access the database.}

\item{database_password}{character defining the password to access the database.}
}
\value{
Returns an object of class \code{sf}, containing all pick-ups,
with geographical location and timestamp.
}
\description{
Retrieves unique bike identification numbers from the JUMP Bikes database,
during a given period. Database credentials are needed for this function.
}
