% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_minmax}
\alias{scale_minmax}
\title{Min/max normalization}
\usage{
scale_minmax(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
Returns a numeric vector.
}
\description{
Normalizes a numeric vector with min/max normalization. Missing values are
allowed.
}
