% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemarea.R
\docType{data}
\name{systemarea}
\alias{systemarea}
\title{System area}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 1 columns.}
\source{
\url{https://us.jump.com/cities/san-francisco/}
}
\usage{
systemarea
}
\description{
The system area of the JUMP Bikes dockless bike sharing system in
San Francisco.
}
\keyword{datasets}
