% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_loop.R
\name{test_seasonality}
\alias{test_seasonality}
\title{Automatic seasonality testing}
\usage{
test_seasonality(data, seasons)
}
\arguments{
\item{data}{object of class \code{dockless_df}.}

\item{seasons}{a list of different seasonal period lengths - or, in the case
of multiple seasonality, combinations of seasonal period lenghts - in time
lags to be tested for. No seasonality should be represented by \code{NULL}.}
}
\value{
Returns a value specifying the length in time lags of the seasonal
period. In the case of multiple seasonality, it returns a vector. In the case
of no seasonality, it returns \code{NULL}.
}
\description{
Finds the seasonal period - or, in the case of multiple seasonality,
combination of seasonal periods - of a \code{dockless_df} time series that
leads to the highest forecast accuracy. Does this with time series
cross-validation.
}
