% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpoints.R
\docType{data}
\name{testpoints}
\alias{testpoints}
\title{Test points}
\format{An object of class sf with POINT geometry, with 500 features
and 3 columns.
\describe{
  \item{time}{timestamp of the test point}
  \item{cluster}{cluster in which the test point lies}
  \item{location}{location of the test point}
}}
\usage{
testpoints
}
\description{
The location-timestamp combinations of the 500 test points.
}
\keyword{datasets}
