% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_fun.R
\name{update_distances}
\alias{update_distances}
\title{Updates a \code{dockless_dfc} object with new data}
\usage{
update_distances(locations, data, database_user, database_password)
}
\arguments{
\item{locations}{object of class \code{sf} with point geometry.}

\item{data}{object of class \code{dockless_dfc} to be updated.}

\item{database_user}{character defining the user name to access the database.}

\item{database_password}{character defining the password to access the database.}
}
\value{
Returns the updated object of class \code{dockless_dfc}.
}
\description{
Queries, for each of the given geographical locations, distances to the
nearest available bike for several timestamps in the past, from the JUMP Bikes
database. Queries only data after the last timestamp of the provided
\code{dockless_dfc} object. Database credentials are needed for this function.
}
