% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_loop.R
\name{usage_intensity}
\alias{usage_intensity}
\title{Usage intensity in grid cells}
\usage{
usage_intensity(usage, grid)
}
\arguments{
\item{usage}{pick-ups, as an object of class \code{sf} with point geometry.}

\item{grid}{grid cells as an object of class \code{sf} with polygon geometry.}
}
\value{
Returns an numeric vector with each element specifying the number
of pick-ups in the grid cell with corresponding index.
}
\description{
Calculates the number of pick-ups in each polygon of an overlaying grid.
}
