% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usagedata_test.R
\docType{data}
\name{usagedata_test}
\alias{usagedata_test}
\title{Usage data for the test period}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 20737 rows and 4 columns.}
\usage{
usagedata_test
}
\description{
The usage data, i.e. all pick-ups, during the test week.
}
\keyword{datasets}
