% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usagedata_train.R
\docType{data}
\name{usagedata_train}
\alias{usagedata_train}
\title{Usage data for the training period}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 56587 rows and 4 columns.}
\usage{
usagedata_train
}
\description{
The usage data, i.e. all pick-ups, during the training period.
}
\keyword{datasets}
