% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weekhour}
\alias{weekhour}
\title{Extract hour of the week}
\usage{
weekhour(x, week_start = 1)
}
\arguments{
\item{x}{timestamp of class \code{POSIXct}.}

\item{week_start}{day on which week starts following ISO conventions -
1 means Monday, 7 means Sunday.}
}
\value{
Returns a numeric value between 0 and 167.
}
\description{
Extracts the hour of the week from a timestamp of class \code{POSIXct}.
}
