% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weighted_centroid}
\alias{weighted_centroid}
\title{Weighted centroid}
\usage{
weighted_centroid(points, weights)
}
\arguments{
\item{points}{object of class \code{sf} with point geometry.}

\item{weights}{vector specifying the weight of each point.}
}
\value{
Returns an object of class \code{sfc_POINT}.
}
\description{
Calculates the weighted centroid of an object of class \code{sf} with
point geometry.
}
