% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_loop.R
\name{create_modelpoints}
\alias{create_modelpoints}
\title{Create model points}
\usage{
create_modelpoints(points, clusters, weights)
}
\arguments{
\item{points}{all grid cell centers as an \code{sf} object with point geometry.}

\item{clusters}{vector specifying to which cluster each grid cell center belongs.}

\item{weights}{vector specifying the usage intensity of each grid cell
corresponding to a grid center.}
}
\value{
Returns an object of class \code{sf} with point geometry.
}
\description{
Creates an object of class \code{sf} containing the geographical locations
for which the forecasting models will be build. The locations are calculated
by taking, per cluster, the centroid of the grid cell centers , weighted by
the usage intensity of the grid cell polygons.
}
