% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_loop.R
\name{create_testpoints}
\alias{create_testpoints}
\title{Create test points}
\usage{
create_testpoints(data, area, clusters, n)
}
\arguments{
\item{data}{the usage data of the test period as object of class \code{sf}
with point geometry.}

\item{area}{the system area of the bike sharing system as object of class
\code{sf}, with polygon geometry.}

\item{clusters}{the geographical outlines of the clusters as object of class
\code{sf}, with polygon geometry.}

\item{n}{the desired number of test points.}
}
\value{
Returns an object of class \code{sf}, with point geometry.
}
\description{
Creates an object of class \code{sf} containing the geographical locations
and timestamps of the test points. All location-timestamp combinations are
sample from the usage data of the test period.
}
