% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{error_cluster}
\alias{error_cluster}
\title{Forecast error per cluster}
\usage{
error_cluster(x, clusters, type)
}
\arguments{
\item{x}{object of class \code{dockless_fcc}.}

\item{clusters}{vector specifying to which cluster each test point belongs.}

\item{type}{one of 'RMSE' or 'MAE'.}
}
\value{
Returns a vector of numeric values.
}
\description{
Calculates either the root mean squared error or the mean absolute
error, per cluster.
}
