% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_fun.R
\name{forecast_lastweek}
\alias{forecast_lastweek}
\title{Forecast last week of a \code{dockless_df} object}
\usage{
forecast_lastweek(data, model, weeks_of_data = NULL)
}
\arguments{
\item{data}{object of class \code{dockless_df}.}

\item{model}{forecasting model. Should be either of class \code{ARIMA}, for
non-seasonal data, or class \code{stlm}, for seasonal data.}

\item{weeks_of_data}{how many weeks of historical data to use for forecasting.
If set to \code{NULL}, all available data is used.}
}
\value{
Returns a list containing seven  - one for each day - objects of
class \code{dockless_fc_df}, which is a data frame containing the observed
values, forecasted values and corresponding prediction intervals of a one-day
forecast for one single time series.
}
\description{
Forecasts the last week of a \code{dockless_df} time series, using a model
generated with either \link{build_single_model} or \link{build_models}.
Each day will be forecasted seperately, with a period of data of given length.
The period of data to forecast with will always end at the start of the day
that is forecasted.
}
