% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{mae}
\alias{mae}
\alias{mae.dockless_fc_df}
\alias{mae.dockless_fc_dfc}
\title{Mean absolute error}
\usage{
mae(x)

\method{mae}{dockless_fc_df}(x)

\method{mae}{dockless_fc_dfc}(x)
}
\arguments{
\item{x}{object either of class \code{dockless_fc_df} or \code{dockless_fc_dfc}.}
}
\value{
Returns a numeric value.
}
\description{
Calculates the mean absolute error of one or more forecasts.
}
