% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{query_distances}
\alias{query_distances}
\title{Query historical distance data}
\usage{
query_distances(locations, from = as.POSIXct("2018-09-10 16:30:00",
  format = "\%Y-\%m-\%d \%H:\%M:\%S", tz = "America/Los_Angeles"),
  to = Sys.time(), database_user, database_password)
}
\arguments{
\item{locations}{object of class \code{sf} with point geometry.}

\item{from}{timestamp of class \code{POSIXct} defining from which timestamp
on data should be queried. Default is set to the first timestamp
in the database (i.e. earlier timestamps than the default are not possible).}

\item{to}{timestamp of class \code{POSIXct} defining until which timestamp
data should be queried. Default is set to the system time.}

\item{database_user}{character defining the user name to access the database.}

\item{database_password}{character defining the password to access the database.}
}
\value{
Returns an object of class \code{dockless_dfc}, which is a collection
of data frames of class \code{dockless_df}. Each \code{dockless_df} object is
a time series of historical distance data for a specific geographical location.
}
\description{
Retrieves, for each of the given geographical locations, distances to the
nearest available bike for several timestamps in the past, from the JUMP Bikes
database. Database credentials are needed for this function.
}
