% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{query_usage}
\alias{query_usage}
\title{Query pick-ups and drop-offs from historical data}
\usage{
query_usage(database_user, database_password)
}
\arguments{
\item{database_user}{character defining the user name to access the database.}

\item{database_password}{character defining the password to access the database.}
}
\value{
Returns an object of class \code{sf}, containing all pick-ups and
drop-offs, with geographical location and timestamp.
}
\description{
Queries individual pick-ups and drop-offs from the complete JUMP Bikes
database. Database credentials are needed for this function.
}
