% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_loop.R
\name{spatial_cluster}
\alias{spatial_cluster}
\title{Spatially constrained hierarchical clustering of a \code{dockless_dfc} object}
\usage{
spatial_cluster(data, grid, area, K, omega = seq(0, 1, 0.1))
}
\arguments{
\item{data}{object of class \code{dockless_dfc}.}

\item{grid}{grid cells as object of class \code{sf} with polygon geometry, in which
each cell refers to the spatial location of each \code{dockless_df} object in
the \code{dockless_dfc}. Must contain a column named \code{intensity}, which provides
the number of pick-ups per grid cill.}

\item{area}{object of class \code{sf} representing the system area.}

\item{K}{vector of integers specifying which values of k, the number of
clusters, should be tested.}

\item{omega}{vector of values specifying which values of alpha, the mixing
parameter, should be tested.}
}
\value{
Returns a list of 2 with one element being a numeric vector of cluster
indices that specifies for each of the given \code{dockless_df} objects to which
cluster it belongs, and the second element being the geographical outines of each
cluster, bundled in an object of class \code{sf} with polygon geometry.
}
\description{
Clusters the \code{dockless_df} objects in a \code{dockless_dfc} with
spatially constrained hierarchical clustering.
}
