% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_loop.R
\name{spatial_cluster}
\alias{spatial_cluster}
\title{Spatially constrained hierarchical clustering of a \code{dockless_dfc} object}
\usage{
spatial_cluster(data, polygons = NULL, k = NULL, alpha = 0)
}
\arguments{
\item{data}{object of class \code{dockless_dfc}.}

\item{polygons}{object of class \code{sf} with polygon geometry, in which each
feature refers to the spatial location of each \code{dockless_df} object in
the \code{dockless_dfc}.}

\item{k}{number of clusters.}

\item{alpha}{value of the mixing parameter alpha, which sets the strength
of the spatial constraint.}
}
\value{
Returns an object of class \code{dockless_clust}, which is a list
containing two elements: 1 - a vector that specifies for each of the given
\code{dockless_df} objects to which cluster it belongs and 2 - object of class
\code{hclust} which describes the tree produced by the clustering process.
}
\description{
Clusters the \code{dockless_df} objects in a \code{dockless_dfc} with
spatially constrained hierarchical clustering.
}
